Use moveToThread for Video Thread
Instantiate a QThread in TTS_GUI and move VideoThread to that thread.

Use Signals/Slots in DirectConnection type to execute slots directly
after a signal is emitted.

Moved setCamera to constructor.

Moved display and saving of video frame to slots:
- process() emits a new frame
- savingVideo(frame) saves frame only if connection established by
saveVideo(bool) which arg is set by signal in TTS_GUI when Start button
is clicked
- emitVideo(frame) emits an image only if connection established by
displayVideo(bool) which is a public slot whose arg set in TTS_GUI